requirements.txt
Django==5.0.8
djangorestframework==3.15.2
psycopg[binary]==3.2.1
dj-database-url==2.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
openpyxl==3.1.5
